Fix vertical splitter in variables pane (#5391)
### Description
Addresses #4720 

Fixes a bug where users were unable to smoothly resize the columns in
the variables pane. This PR changes how the vertical splitter handles
the events used to resize columns. The vertical splitter component would
re-render itself upon a resize which would cause relevant state
information to be reset. We now directly interact with the DOM element
to avoid this issue.

### QA Notes

- [ ] Verify the vertical splitter respects changes made to the sash
size from Positron settings
- [ ] Verify the vertical splitter in the data explorer panel behavior
has not changed and matches behavior in production version of Positron
- [ ] Verify the vertical splitter in the variables panel can be dragged
around (and still respects the min and max widths)

### Screenshots


https://github.com/user-attachments/assets/98e7b29f-0465-40ec-a3b4-031853e38688