Fix notebook restart always timing out (#5461)
Addresses #5460. I think there's been a change somewhere that causes the
session to be "ready" when `restartSession()` resolves, so just moved
the promise creation to before the call which is probably safer anyway.

The `NotebookSessionService` will be deprecated soon; this is a quickfix
until then.

### QA Notes

Restarting a notebook should not time out.