e2e-test: add savePlot (#5479)
### Intent

The goal of this PR was to improve the test by removing a hardcoded wait
that was previously required for the test to work and pass. I was able
to make improvements to the test and it runs/passes but occasionally I
kept running into a flake of RPC timeout on CI runs for the python plot
test. Maybe that was the original issue? (The in-line comment didn't
say). Adding the wait back in does seem to remove the flake, so maybe
that was it after-all. Regardless, there are still improvements worth
merging in.

### Approach

A new method, savePlot, was introduced to encapsulate and enhance the
logic for saving a plot to a selected file type. Additionally, some
before/after hooks were cleaned up for better readability.

### QA Notes
* Confirmed plots test [PASS on
Windows](https://d38p2avprg8il3.cloudfront.net/playwright-report-12008951342/index.html#?q=plots).
* I did not see an open issue for the RPC timeout? If we don't have one,
I can add it and tag it to this test.