Remove second copy of `positron.r.extraArguments` configuration (#5504)
I noticed this while putting something else behind a feature flag. We've
got 2 complete copies of the configuration for
`positron.r.extraArguments`. It was introduced in #3950.

I had to decide which section to leave this in and decided it was
advanced, but this is not a hill I will die on. I'm fine with keeping it
in the not-advanced settings and removing from advanced, either way.