Use new Quarto context keys in keyboard shortcuts (#5451)
Addresses #1955 together with
https://github.com/quarto-dev/quarto/pull/608

The Quarto PR provides new context keys for the main language of a
Quarto document (for example, R or Python) and we can consume those to
provide keyboard shortcuts.

There are really only a couple of R ones that we want right now which
means that there is no real change here in an R `.qmd` in behavior
compared to what is in the RStudio Keymap. 🙈 However, I think this is
still worth getting in since we've gone to the trouble of figuring it
out, it lets us offer these without the RStudio Keymap being on, we can
offer them _only_ in R `.qmd` files (right now, these also show up in
Python `.qmd` files if you have the RStudio Keymap on), and we can use
this infrastructure in the future.

I did remove these from the RStudio Keymap just to clean things up, but
it wouldn't hurt _a lot_ to keep them, if there is some reason I am not
thinking of? If we do want to remove them, we'll need to do a Quarto
release and update the bundled Quarto VS Code extension version in this
PR to keep these keybindings functional.

### QA Notes

With the new version of the Quarto extension (updated here), you can use
the keyboard shortcuts for the pipe and assignment operator in R Quarto
documents but not Python ones:

- <kbd>Alt</kbd>+<kbd>-</kbd> to get you `<-`
- <kbd>Cmd/Ctrl</kbd>+<kbd>Shift</kbd>+<kbd>M</kbd> to get you `|>`