Fix R package installation commands after multiple restarts (#5579)
This change addresses the remainder of issue
https://github.com/posit-dev/positron/issues/5532.

The underlying problem turned out to be that in some circumstances
(depending on timing) it is possible to get two UI comms to open at the
same time. This happens when the _Ready_ status arrives multiple times.
Since the _Ready_ status triggers the creation of the UI comm, and
creating the comm is an async operation, receiving it multiple times can
result in multiple in-flight requests to create the comm.

The fix has a few parts:

1. The fact that the _Ready_ status event happens multiple times can be
traced upstream to the supervisor, which can mark the kernel Ready after
getting the kernel info _and_ after the initial heartbeat. This fix
updates to version 0.21 of the supervisor, which doesn't have this bug.
2. Even if the runtime fires multiple _Ready_ events, we should not
create multiple UI comms. This fix adds re-entrancy guards to the part
of the system that creates the comm, so that if we are asked to create
one when one is already active, or we are in the process of creating
one, we don't wind up with two.
3. Finally, I've done some light refactoring of the active session state
to reduce the size of the runtime session service, and added a test for
this bug.

### QA Notes

In order to test this, you'll want to be sure that the scenario in #5532
works at least 5 times in a row. I also recommend enabling the Runtime
Sessions debug panel so you can check to see that the correct number of
UI comms are in play as you test.