Simplify logic for switching runtimes for a notebook (#5558)
A little refactor to simplify the logic for switching runtimes for a
notebook.

When a user switches from one Positron notebook controller to another,
two events are fired: one with `selected === true` and one with
`selected === false`. The order of those events is not guaranteed, so we
have to be careful about the order we request to shutdown the previous
runtime and start the new one.

We previously worked around that by retrying the start request, but it's
much simpler to shutdown any existing runtimes in the `selected ===
true` case.

### QA Notes

In a notebook, if you change the selected runtime a bunch of times you
should be able to execute code in any selected runtime and there should
not be any error notifications or console logs.