Python: Do not execute Ibis expressions when ibis.options.interactive is True (#5625)
Addresses #5499 by adding a custom inspector for Ibis expressions. This
is very basic, and per #5573 should perhaps live eventually in Ibis
itself.

Ibis is a bit unusual in that its interactive mode causes computation to
be executed when running the `__repr__` method, for nice interactivity
in the console and in Jupyter notebooks. So here we avoid running the
`__repr__` method so we don't accidentally fire off a BigQuery,
Snowflake, or other query which might have unwanted costs or side
effects.

There is a unit test -- Ibis with DuckDB is a minor dependency to pull
in relative to the rest of our test dependencies so I do not think this
is too onerous.