Data Explorer: Relax DuckDB post-query checks to allow queries past the end of tables (#5644)
Addresses #5552. The frontend in some of its UI calculations requests
the first 10 rows of data, but there may not be 10 rows available. It
might be good to improve the UI logic, but in the meantime there is no
reason to error when a query requests beyond the end of the table.