Do not error in comm code generator when 'name' property present in object that has enum (#5645)
Addresses #2139. This isn't causing active harm, but a sharp edge that
will hit comm developers with a confusing error message when an object
property is called "name" and there is another property that is an enum.
I tested this out with the example from the linked issue and it works
correctly. I also checked that this change does not impact any of the
generated comm code.