Data Explorer: When a table has row labels, do not return an incorrect row label while RPC is pending (#5654)
Addresses issue described in #5230. When the row label cache has not
been updated yet (the backend getRowLabels RPC is pending), the data
cache was falling back on returning the row index, causing the row label
to appear to be temporarily wrong. There are videos in #5230 showing
exactly this.

Now the string `...` is showed momentarily to indicate that the row
labels are still loading. This seems less misleading:

https://github.com/user-attachments/assets/be14c456-ab0a-42b1-8027-ec4624b24e90