Improve UX when no runtimes found (#5671)
### Description
Addresses #3468

This addresses a niche UX scenario: a user installs Positron before
completing the [machine
prerequisites](https://positron.posit.co/start.html#machine-prerequisites)
i.e. there is no R or Python installed on the machine.

The solution is quite minimal since the overall runtime startup
experience will be going through enhancements soon. A notification is
displayed to the user letting them know that no runtimes were found.
This notification is shown after the runtime discovery phase is
complete.

<img width="1342" alt="Screenshot 2024-12-09 at 9 41 33 AM"
src="https://github.com/user-attachments/assets/65e7f20d-4497-46fd-85e5-ef361f5809c7">

### QA Notes

In order to test this we'll need to install Positron in an environment
that does not have R or Python installed. This should kick off a VS Code
notification in the bottom right which lets a user know the problem and
a link to the positron wiki for resolving the issue.

---------

Signed-off-by: Dhruvi Sompura <dhruvi.sompura@posit.co>
Co-authored-by: Julia Silge <julia.silge@gmail.com>