Add configurable session timeouts for supervisor (#5693)
Addresses #5620 by adding an option to specify the number of seconds to
wait for the kernel to connect before giving up. This change is mostly
on the supervisor side; the client changes here just add the new option
and pass it along to the supervisor.

<img width="507" alt="image"
src="https://github.com/user-attachments/assets/89ae4191-3865-4637-b301-e5660e61cd4d">

The default was 20 (hardcoded) and is now 30 (configurable).

While I was updating the API, I also added some methods/types that will
later support #5226 (but aren't currently implemented on the client or
server side).

### QA Notes

A really easy way to test this is to set the timeout at 1 second, since
most kernels do not start up that quickly.