Notebook variables pane sync (#5583)
Addresses #2452

This PR adds the ability for the variables pane to "follow" the user
around as they work across different sessions (particularly useful for
notebooks that have their own sessions.)

It works by automatically switching the variables pane in the following
circumstances
- If the active editor tab is changed
- If changed to a notebook, will switch variables to that notesbook
session's variables
- If changed to a non-notebook file such as a script, will switch the
variables to the last-active console session.
- If code is run in the console the session backing the console's
variables will be focused.


https://github.com/user-attachments/assets/7fae8c0d-eea4-4cd4-a664-ee7cab902674

All of this behavior can be turned off by setting the
`positron.variables.followMode` setting to false.
(`src/vs/workbench/services/positronVariables/common/positronVariables.contribution.ts`)

### Annoyances
- Currently there's no real way to correlate a script like a `.R` or
`.py` script with a session. In this case we go back to the most recent
console session, but that could be confusing if you clicked onto an R
session and your most recent console was a python session. I left this
behavior as is because it's what _currently_ happens and it feels like
the upcoming multi-session work will formalize behavior around this.

### Code Refactoring:

* Renamed the method `setActivePositronVariablesInstance` to
`_setActivePositronVariablesInstance` for consistency with other private
methods.
(`src/vs/workbench/services/positronVariables/common/positronVariablesService.ts`)


## QA notes:

Should be able to fiddle between different notebooks and consoles and
the variables pane should go to where you expect it to.