Data Explorer: Visually display leading/trailing whitespace and empty strings in data values (#5652)
Addresses #3067. Currently in the data explorer, the strings `""`, `"
"`, and `" "` cannot be visually distinguished because of the absence of
quotations.

There are a few approaches we could take:

1. Format with quotations (I believe RStudio does this)
2. Provide a special display for empty strings (this is what's described
as #3067)
3. Replace whitespace with a unicode character to indicate the presence
of leading/trailing whitespace characters

This PR does the latter two, using a mid-dot as the unicode
representation and the special string `<empty>` for empty strings.