Package 'positron-proxy' and 'positron-duckdb' dependencies (#5756)
This change fixes the Positron Proxy and Positron-DuckDB
(CSV/TSV/Parquet viewer) extension bundling in release builds. An
upstream change caused dependencies of built-in extensions to be
ignored; this is mostly fine but a couple of our built-in extensions
have third-party dependencies that need to be processed.

The fix is to use the NPM dependency detection strategy on these
specific extensions. We can't just use this strategy on _all_ extensions
since it does not work on extensions that have dependencies that are
externally supplied (which is the case for several Microsoft and
Positron extensions).

Additional fixes:
- Fixes loading the `help.html` file in the Help pane in release builds,
and makes the error visible if we can't.
- Serializes Webpack extension building to avoid running out of file
handles on Windows (`EMFILE`).

Addresses https://github.com/posit-dev/positron/issues/5742
Addresses https://github.com/posit-dev/positron/issues/5743



e2e: @:data-explorer

### QA Notes

This problem only reproduces in release builds.