Fix Python interpreter not found error (#5793)
Attempts to address #5730. My hunch is that `getInterpreters()` doesn't
wait for the Python extension to complete its discovery process, so it's
possible to get there before our runtime's interpreter is discovered.

The whole point of restoring a workspace-affiliated runtime is to skip
discovery. It looks like `getInterpreterDetails()` is a better fit,
since (I think) it "resolves" an interpreter path to a
`PythonEnvironment` independent of discovery.

### QA Notes

E2E tests C609619 and C609617 should pass in CI:

* Passing E2E run:
https://github.com/posit-dev/positron/actions/runs/12387951151/job/34578222238
* 10x run of both failing tests:
https://d38p2avprg8il3.cloudfront.net/playwright-report-12389107722-6803/index.html.
There was a single failure, but that looks like an unrelated issue.

Python and Reticulate runtimes should continue to function as expected.