Ensure the supervisor can start even when `http_proxy` is set (#5825)
We have had several user reports of the kernel supervisor failing to
start because of HTTP proxy settings (specifically via the `http_proxy`
environment variable). This happens because the Node request library we
use reads the `http_proxy` environment variable and always attempts to
use the proxy (even, surprisingly enough, for local connections).

The fix here is -- when `http_proxy` is set -- to amend the `no_proxy`
environment variable so that the proxy is not used when connecting to
the supervisor.

Partially addresses https://github.com/posit-dev/positron/issues/5481.

### QA Notes

It can be a little tricky to set environment variables for Positron. The
easiest way is probably to set them in a terminal and then launch
Positron from that terminal.