Make output channel names consistent (#5833)
### Description
Addresses #1707

@:output

We've received feedback that the output channel list is confusing and
not easy to parse. A lot of the channel names are prefixed with
"Positron" which is redundant. The formatting of the names for the
language channels is not consistent and adds friction to the UX when
scanning the list.

The proposed change here does some tidying 🧹 of the names to improve the
UX:
* The "Positron" prefix has been removed
* The language specific channel names have a consistent formatting which
groups them together in the select
* Updates ambiguous output channel names

### Screenshots

<img width="1610" alt="Screenshot 2024-12-19 at 10 20 58 AM"
src="https://github.com/user-attachments/assets/4ddad90c-62c6-4044-bf51-f977436ea8cc"
/>

<img width="1610" alt="Screenshot 2024-12-19 at 10 25 38 AM"
src="https://github.com/user-attachments/assets/fee1ccde-e90d-4162-800e-baca2a95ae41"
/>