Data Explorer: Visually distinguish column names with leading/trailing whitespace or empty string (#5653)
Aims to address #3084, #3089. Related to #5652, we also cannot
distinguish empty strings or strings with leading whitespace in column
names both in the data grid headers and the summary panel.

This PR applies the same approach we used for data cells to format
column names in the grid headers and the summary pane.